requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0

# Async support
asgiref==3.9.2

billiard==4.2.2
cachetools==6.2.0
celery==5.5.3
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==5.2.6
django-bootstrap5==25.2
django-countries==7.6.1
django-timezone-field==7.1
djangorestframework==3.16.1
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
jiter==0.11.0
kombu==5.5.4
msgpack==1.1.1
oauthlib==3.3.1
openai==2.0.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
